Adding default handling of kernel_file_name
Currently the craycli is passing a value of `vmlinuz` which is a
default for x86 arch. IMS needs to support the passing of both
`Image` for aarch64, and `vmlinuz` for x86. Changes the behavior such that if input to the API was either a null or empty or a user error, the file name will be determined by the arch type defined as a last resort.

Additionally was an error in the V3 tests where the tests were testing V2 endpoints.

Added V3 tests that reflect the changes.

Fixed flask test server holding too many connections locally when running tests. Separated out new V3 tests to solve this.